- Homework 26.8 is finished; - Added possibility to share content;
